pyproject.toml
[project]
name = "road-damage-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "roboflow>=1.2.16",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.23",
    "uvicorn>=0.42.0",
]
